api/requirements.txt
flask==3.0.3
flask-cors==4.0.0
pytest==8.1.1
requests==2.31.0
kafka-python==2.0.2
